Add types
This change introduces a '.d.ts' file. The change doesn't alter the source
in any way and does not introduce Typescript to the project. However, this
file will allow this package to be used by other projects using Typescript.

The webpack copy plugin was added as well to copy the file to the dist/
directory.

The types in the file were sourced almost entirely from the API Reference:
https://geotab.github.io/sdk/software/api/reference/